requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pymongo>=4.6.0
motor>=3.3.0
requests>=2.31.0
twilio>=8.10.0
tensorflow>=2.15.0
opencv-python>=4.8.0
numpy>=1.24.0
